Add bottom margin to prevent iPhone bottom navbar
When using the attendance screens, an iPhone 12 mini is very small and the "new
attendance" form ends up touching the bottom of the screen. Trying to use the
buttons at the bottom of the screen ends up showing the navbar, instead of
hitting the button we were trying to use. Instead, we had to tap the button,
scroll the page, then hit the button.

Added a margin using an expedient shortcut, instead of using nice Bootstrap.
Yeah, yeah, I know, but it works!